requirements.txt
# AI Agent Education Platform Requirements
# Updated: 2024-09-10

# Core Framework
fastapi==0.115.14
uvicorn[standard]==0.35.0

# Database
SQLAlchemy>=2.0.0,<3.0
alembic==1.16.2
# PostgreSQL support (primary database)
psycopg2-binary==2.9.10
# SQLite support (development only - optional)
# sqlite3 is included with Python standard library

# Authentication & Security
passlib[bcrypt]==1.7.4
bcrypt>=3.2.0,<4.0.0
python-jose[cryptography]==3.5.0
python-multipart==0.0.20
# OAuth dependencies
authlib==1.3.0
httpx>=0.28.1
google-auth>=2.0.0
PyJWT>=2.0.0

# AI Services
openai>=1.97.1
anthropic==0.55.0

# LangChain Framework
langchain==0.3.7
langchain-openai==0.2.8
langchain-community==0.3.7
langchain-postgres==0.0.12
langchain-experimental==0.3.4

# Vector Database & Embeddings
pgvector>=0.2.5,<0.3.0
sentence-transformers==3.1.1
faiss-cpu==1.12.0

# Caching & Session Management
redis==5.2.0
python-memcached==1.62

# Additional LangChain Tools
langchain-text-splitters==0.3.2

# HTTP & Async
aiohttp==3.10.11
aiofiles==24.1.0

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.11.7
pdfminer.six==20250506

# Configuration & Environment
python-dotenv==1.1.1
pydantic==2.11.7
pydantic-settings==2.6.1

# Data Processing
PyYAML==6.0.2

# Development & Testing (optional)
pytest==8.4.1
pytest-asyncio==0.24.0

# Database Admin Interface (optional)
flask==3.0.3
flask-admin==1.6.1

# Additional utilities
pathlib2==2.3.7
typing-extensions==4.12.2
schedule==1.2.2

# Scientific Computing
numpy>=1.26.0,<2.0.0  # Used in vector_store.py for array operations, compatible with LangChain

# Note: The following are built-in Python modules and don't need to be installed:
# hashlib, base64, pickle, asyncio, json, datetime, time, os, sys, 
# logging, subprocess, re, string, unicodedata 

